Another change to fix the 32 bit / 64 bit compatibility

+++ src/larcv3/core/dataformat/DataFormatTypes.cxx
@@ -71,7 +71,7 @@ namespace larcv3{
     herr_t status;
     datatype = H5Tcreate (H5T_COMPOUND, sizeof (Extents_t));
     status = H5Tinsert (datatype, "first",
-                HOFFSET (Extents_t, first), larcv3::get_datatype<unsigned int>());
+                HOFFSET (Extents_t, first), larcv3::get_datatype<unsigned long long int>());
     status = H5Tinsert (datatype, "N", 
                 HOFFSET (Extents_t, n),     larcv3::get_datatype<unsigned int>());
     return datatype;
@@ -84,7 +84,7 @@ namespace larcv3{
     herr_t status;
     datatype = H5Tcreate (H5T_COMPOUND, sizeof (IDExtents_t));
     status = H5Tinsert (datatype, "first",
-                HOFFSET (IDExtents_t, first), larcv3::get_datatype<unsigned int>());
+                HOFFSET (IDExtents_t, first), larcv3::get_datatype<unsigned long long int>());
     status = H5Tinsert (datatype, "N", 
                 HOFFSET (IDExtents_t, n),     larcv3::get_datatype<unsigned int>());
     status = H5Tinsert (datatype, "ID",
